# docs/swagger.yaml
basePath: /
definitions:
...
  input.CreateVehicle:
    description: Dados necessários para criar um novo veículo para o usuário autenticado
    properties:
      brand:
        description: Marca do veículo
        example: Honda
        type: string
      mainPhotoContentType:
        description: 'Tipo de conteúdo da foto principal (opcional: image/jpeg, image/png,
          image/webp)'
        enum:
        - image/jpeg
        - image/png
        - image/webp
        example: image/jpeg
        type: string
      model:
        description: Modelo do veículo
        example: CB 600F
        type: string
      nickname:
        description: Apelido do veículo
        example: Minha Moto
        type: string
      version:
        description: Versão do veículo (opcional)
        example: Hornet
        type: string
      year:
        description: Ano do veículo (deve ser maior que 0)
        example: 2020
        type: integer
    required:
    - brand
    - model
    - nickname
    - year
    type: object
...
  input.RequestPhotoUpsert:
    description: Dados para solicitar URL pré-assinada para upload de foto principal
      ou de feed do veículo
    properties:
      contentType:
        description: Tipo de conteúdo da foto (image/jpeg, image/png, image/webp)
        enum:
        - image/jpeg
        - image/png
        - image/webp
        example: image/jpeg
        type: string
      photoType:
        description: Tipo de foto (main ou feed)
        enum:
        - main
        - feed
        example: main
        type: string
    required:
    - contentType
    - photoType
    type: object
...
  input.UpdateVehicleInfo:
    description: Dados opcionais para atualizar informações de um veículo existente
    properties:
      brand:
        description: Marca do veículo (opcional)
        example: Yamaha
        type: string
      model:
        description: Modelo do veículo (opcional)
        example: MT-07
        type: string
      nickname:
        description: Apelido do veículo (opcional)
        example: Minha Moto Atualizada
        type: string
      version:
        description: Versão do veículo (opcional)
        example: ABS
        type: string
      year:
        description: Ano do veículo (opcional, deve ser maior que 0)
        example: 2021
        type: integer
    type: object
...
  response.Photo:
    description: Informações sobre uma foto adicional do veículo
    properties:
      id:
        description: ID da foto
        example: 456
        type: integer
      reference:
        description: URL da foto
        example: https://example.com/vehicles/123/photos/456.jpg
        type: string
    type: object
...
  response.PresignedPhotoInfo:
    description: URL pré-assinada e chave para upload de foto do veículo
    properties:
      objectKey:
        description: Chave da foto no storage
        example: vehicles/123/photo.jpg
        type: string
      presignedVehiclePhotoUrl:
        description: URL pré-assinada para upload da foto do veículo
        example: https://s3.amazonaws.com/bucket/vehicles/123/photo.jpg?X-Amz-Signature=...
        type: string
      vehicleId:
        description: ID do veículo
        example: 123
        type: integer
    type: object
...
  response.Vehicle:
    description: Informações completas de um veículo retornado na busca/listagem
    properties:
      brand:
        description: Marca do veículo
        example: Honda
        type: string
      createdAt:
        description: Data de criação do veículo
        example: "2024-01-15T10:30:00Z"
        type: string
      id:
        description: ID do veículo
        example: 123
        type: integer
      mainPhotoUrl:
        description: URL da foto principal do veículo (opcional)
        example: https://example.com/vehicles/123/main.jpg
        type: string
      model:
        description: Modelo do veículo
        example: Civic
        type: string
      nickname:
        description: Apelido do veículo
        example: Ligeirinho
        type: string
      photos:
        description: Lista de fotos adicionais do veículo
        items:
          $ref: '#/definitions/response.Photo'
        type: array
      version:
        description: Versão do veículo (opcional)
        example: Hatch EG8
        type: string
      year:
        description: Ano do veículo
        example: 1994
        type: integer
    type: object
...
  src_core_vehicle_controller.ConfirmNewPhotoRequest:
    properties:
      key:
        type: string
    required:
    - key
    type: object
  src_core_vehicle_controller.ErrorMessageResponse:
    properties:
      message:
        type: string
    type: object
  src_core_vehicle_controller.ValidationErrorResponse:
    properties:
      errors:
        additionalProperties:
          type: string
        type: object
    type: object
host: localhost:3000
info:
  contact: {}
  description: API do Revit integrada com Swagger via Fiber.
  title: Revit API
  version: "1.0"
paths:
...
  /api/vehicle:
    get:
      consumes:
      - application/json
      description: Retorna a lista de veículos do usuário autenticado
      parameters:
      - description: Limite de resultados por página (opcional, máximo 20)
        example: 10
        in: query
        maximum: 20
        name: limit
        type: integer
      - description: Apelido do veículo para busca (opcional)
        example: Minha Moto
        in: query
        name: nickname
        type: string
      - description: Número da página (opcional, deve ser maior que 0)
        example: 1
        in: query
        name: page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Lista de veículos
          schema:
            items:
              $ref: '#/definitions/response.Vehicle'
            type: array
        "400":
          description: Erro na validação dos parâmetros
          schema:
            $ref: '#/definitions/src_core_vehicle_controller.ValidationErrorResponse'
        "401":
          description: Token inválido ou expirado
          schema:
            $ref: '#/definitions/src_core_vehicle_controller.ErrorMessageResponse'
      security:
      - BearerAuth: []
      summary: Listar veículos
      tags:
      - Vehicles
    post:
      consumes:
      - application/json
      description: Cria um novo veículo para o usuário autenticado
      parameters:
      - description: Dados do veículo
        in: body
        name: vehicle
        required: true
        schema:
          $ref: '#/definitions/input.CreateVehicle'
      produces:
      - application/json
      responses:
        "201":
          description: URL pré-assinada para upload da foto principal
          schema:
            $ref: '#/definitions/response.PresignedPhotoInfo'
        "400":
          description: Erro na validação dos dados
          schema:
            $ref: '#/definitions/src_core_vehicle_controller.ValidationErrorResponse'
        "401":
          description: Token inválido ou expirado
          schema:
            $ref: '#/definitions/src_core_vehicle_controller.ErrorMessageResponse'
      security:
      - BearerAuth: []
      summary: Criar veículo
      tags:
      - Vehicles
  /api/vehicle/{vehicleId}:
    delete:
      consumes:
      - application/json
      description: Remove um veículo do usuário autenticado
      parameters:
      - description: ID do veículo
        in: path
        name: vehicleId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: Veículo deletado com sucesso
        "400":
          description: ID inválido
          schema:
            $ref: '#/definitions/src_core_vehicle_controller.ErrorMessageResponse'
        "401":
          description: Token inválido ou expirado
          schema:
            $ref: '#/definitions/src_core_vehicle_controller.ErrorMessageResponse'
        "403":
          description: Sem permissão para deletar o veículo
          schema:
            $ref: '#/definitions/src_core_vehicle_controller.ErrorMessageResponse'
        "404":
          description: Veículo não encontrado
          schema:
            $ref: '#/definitions/src_core_vehicle_controller.ErrorMessageResponse'
      security:
      - BearerAuth: []
      summary: Deletar veículo
      tags:
      - Vehicles
    patch:
      consumes:
      - application/json
      description: Atualiza as informações de um veículo
      parameters:
      - description: ID do veículo
        in: path
        name: vehicleId
        required: true
        type: integer
      - description: Dados do veículo para atualização
        in: body
        name: vehicle
        required: true
        schema:
          $ref: '#/definitions/input.UpdateVehicleInfo'
      produces:
      - application/json
      responses:
        "204":
          description: Veículo atualizado com sucesso
        "400":
          description: Erro na validação dos dados
          schema:
            $ref: '#/definitions/src_core_vehicle_controller.ValidationErrorResponse'
        "401":
          description: Token inválido ou expirado
          schema:
            $ref: '#/definitions/src_core_vehicle_controller.ErrorMessageResponse'
        "403":
          description: Sem permissão para atualizar o veículo
          schema:
            $ref: '#/definitions/src_core_vehicle_controller.ErrorMessageResponse'
        "404":
          description: Veículo não encontrado
          schema:
            $ref: '#/definitions/src_core_vehicle_controller.ErrorMessageResponse'
      security:
      - BearerAuth: []
      summary: Atualizar informações do veículo
      tags:
      - Vehicles
  /api/vehicle/{vehicleId}/photo/{photoId}:
    delete:
      consumes:
      - application/json
      description: Remove uma foto específica de um veículo
      parameters:
      - description: ID do veículo
        in: path
        name: vehicleId
        required: true
        type: integer
      - description: ID da foto
        in: path
        name: photoId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: Foto removida com sucesso
        "400":
          description: ID inválido
          schema:
            $ref: '#/definitions/src_core_vehicle_controller.ErrorMessageResponse'
        "401":
          description: Token inválido ou expirado
          schema:
            $ref: '#/definitions/src_core_vehicle_controller.ErrorMessageResponse'
        "403":
          description: Sem permissão para atualizar o veículo
          schema:
            $ref: '#/definitions/src_core_vehicle_controller.ErrorMessageResponse'
        "404":
          description: Veículo ou foto não encontrado
          schema:
            $ref: '#/definitions/src_core_vehicle_controller.ErrorMessageResponse'
      security:
      - BearerAuth: []
      summary: Remover foto do veículo
      tags:
      - Vehicles
  /api/vehicle/main-photo/{vehicleId}:
    delete:
      consumes:
      - application/json
      description: Remove a foto principal de um veículo
      parameters:
      - description: ID do veículo
        in: path
        name: vehicleId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: Foto principal removida com sucesso
        "400":
          description: ID inválido
          schema:
            $ref: '#/definitions/src_core_vehicle_controller.ErrorMessageResponse'
        "401":
          description: Token inválido ou expirado
          schema:
            $ref: '#/definitions/src_core_vehicle_controller.ErrorMessageResponse'
        "403":
          description: Sem permissão para atualizar o veículo
          schema:
            $ref: '#/definitions/src_core_vehicle_controller.ErrorMessageResponse'
        "404":
          description: Veículo não encontrado
          schema:
            $ref: '#/definitions/src_core_vehicle_controller.ErrorMessageResponse'
      security:
      - BearerAuth: []
      summary: Remover foto principal do veículo
      tags:
      - Vehicles
  /api/vehicle/photo/{vehicleId}:
    patch:
      consumes:
      - application/json
      description: Confirma o upload de foto após upload na URL pré-assinada
      parameters:
      - description: ID do veículo
        in: path
        name: vehicleId
        required: true
        type: integer
      - description: Confirmação do upload
        in: body
        name: confirmation
        required: true
        schema:
          $ref: '#/definitions/src_core_vehicle_controller.ConfirmNewPhotoRequest'
      produces:
      - application/json
      responses:
        "204":
          description: Foto confirmada com sucesso
        "400":
          description: Erro na validação dos dados
          schema:
            $ref: '#/definitions/src_core_vehicle_controller.ValidationErrorResponse'
        "401":
          description: Token inválido ou expirado
          schema:
            $ref: '#/definitions/src_core_vehicle_controller.ErrorMessageResponse'
        "403":
          description: Sem permissão para atualizar o veículo
          schema:
            $ref: '#/definitions/src_core_vehicle_controller.ErrorMessageResponse'
        "404":
          description: Veículo não encontrado
          schema:
            $ref: '#/definitions/src_core_vehicle_controller.ErrorMessageResponse'
      security:
      - BearerAuth: []
      summary: Confirmar upload de foto do veículo
      tags:
      - Vehicles
    post:
      consumes:
      - application/json
      description: Solicita uma URL pré-assinada para upload de foto do veículo
      parameters:
      - description: ID do veículo
        in: path
        name: vehicleId
        required: true
        type: integer
      - description: Dados da requisição de foto
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/input.RequestPhotoUpsert'
      produces:
      - application/json
      responses:
        "200":
          description: URL pré-assinada para upload
          schema:
            $ref: '#/definitions/response.PresignedPhotoInfo'
        "400":
          description: Erro na validação dos dados
          schema:
            $ref: '#/definitions/src_core_vehicle_controller.ValidationErrorResponse'
        "401":
          description: Token inválido ou expirado
          schema:
            $ref: '#/definitions/src_core_vehicle_controller.ErrorMessageResponse'
        "403":
          description: Sem permissão para atualizar o veículo
          schema:
            $ref: '#/definitions/src_core_vehicle_controller.ErrorMessageResponse'
        "404":
          description: Veículo não encontrado
          schema:
            $ref: '#/definitions/src_core_vehicle_controller.ErrorMessageResponse'
      security:
      - BearerAuth: []
      summary: Solicitar upload de foto do veículo
      tags:
      - Vehicles
securityDefinitions:
  BearerAuth:
    description: 'Digite "Bearer " seguido do token JWT. Exemplo: "Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"